pyproject.toml
[project]
name = "goit-web-hw-13"
version = "0.1.0"
description = ""
authors = [
    {name = "Viktor Chayun",email = "VChayun@gmail.com"}
]
readme = "README.md"
requires-python = "3.11.4"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pycryptodome (>=3.22.0,<4.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "cloudinary (>=1.44.0,<2.0.0)",
    "django (>=5.2.1,<6.0.0)",
    "scrapy (>=2.13.0,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "pillow (>=11.2.1,<12.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
